CentOS7 实时性配置

+++ linux-5.0/copy.sh
@@ -12,6 +12,10 @@ dst_dir="../test/linux-5.0/"
 if [ -f $file ]; then
 	swapfile=$(echo $file | sed "s/\//-/g")
 	echo Copy $swapfile to $dst_dir
+	if [ -f $dst_dir/${swapfile:1} ]; then
+		echo Exist: $dst_dir/${swapfile:1}
+		exit
+	fi
 	cp $file $dst_dir/$swapfile
 else
 	echo "Wrong file TYPE."
